Fix logout.
