+++ app/src/main/java/chat/rocket/android/chatrooms/di/ChatRoomsFragmentModule.kt
@@ -3,11 +3,9 @@ package chat.rocket.android.chatrooms.di
 import android.arch.lifecycle.LifecycleOwner
 import chat.rocket.android.chatrooms.presentation.ChatRoomsView
 import chat.rocket.android.chatrooms.ui.ChatRoomsFragment
-import chat.rocket.android.core.lifecycle.CancelStrategy
 import chat.rocket.android.dagger.scope.PerFragment
 import dagger.Module
 import dagger.Provides
-import kotlinx.coroutines.experimental.Job
 
 @Module
 @PerFragment
@@ -21,10 +19,5 @@ class ChatRoomsFragmentModule {
     @Provides
     fun provideLifecycleOwner(frag: ChatRoomsFragment): LifecycleOwner {
         return frag
-    }
-
-    @Provides
-    fun provideCancelStrategy(owner: LifecycleOwner, jobs: Job): CancelStrategy {
-        return CancelStrategy(owner, jobs)
     }
 }
