--- app/src/main/java/chat/rocket/android/main/di/MainActivityModule.kt
@@ -1,23 +1,0 @@
-package chat.rocket.android.main.di
-
-import android.arch.lifecycle.LifecycleOwner
-import chat.rocket.android.core.lifecycle.CancelStrategy
-import chat.rocket.android.main.presentation.MainView
-import chat.rocket.android.main.ui.MainActivity
-import dagger.Module
-import dagger.Provides
-import kotlinx.coroutines.experimental.Job
-
-@Module
-class MainActivityModule {
-
-//    @Provides
-//    fun provideMainView(activity: MainActivity): MainView = activity
-
-    @Provides
-    fun provideLifecycleOwner(activity: MainActivity): LifecycleOwner = activity
-
-    @Provides
-    fun provideCancelStrategy(owner: LifecycleOwner, jobs: Job): CancelStrategy = CancelStrategy(owner, jobs)
-
-}
