--- app/src/main/java/chat/rocket/android/main/di/MainActivityProvider.kt
@@ -1,12 +1,0 @@
-package chat.rocket.android.main.di
-
-import chat.rocket.android.main.ui.MainActivity
-import dagger.Module
-import dagger.android.ContributesAndroidInjector
-
-@Module
-abstract class MainActivityProvider {
-
-    @ContributesAndroidInjector(modules = [MainActivityModule::class])
-    abstract fun provideMainActivity(): MainActivity
-}
