+++ app/src/main/java/chat/rocket/android/main/di/MainModule.kt
@@ -1,11 +1,15 @@
 package chat.rocket.android.main.di
 
+import android.arch.lifecycle.LifecycleOwner
 import android.content.Context
+import chat.rocket.android.core.lifecycle.CancelStrategy
 import chat.rocket.android.dagger.scope.PerActivity
 import chat.rocket.android.main.presentation.MainNavigator
+import chat.rocket.android.main.presentation.MainView
 import chat.rocket.android.main.ui.MainActivity
 import dagger.Module
 import dagger.Provides
+import kotlinx.coroutines.experimental.Job
 
 @Module
 class MainModule {
@@ -13,4 +17,13 @@ class MainModule {
     @Provides
     @PerActivity
     fun provideMainNavigator(activity: MainActivity, context: Context) = MainNavigator(activity, context)
+
+    @Provides
+    fun provideMainView(activity: MainActivity): MainView = activity
+
+    @Provides
+    fun provideLifecycleOwner(activity: MainActivity): LifecycleOwner = activity
+
+    @Provides
+    fun provideCancelStrategy(owner: LifecycleOwner, jobs: Job): CancelStrategy = CancelStrategy(owner, jobs)
 }
