+++ app/src/main/java/chat/rocket/android/main/presentation/MainPresenter.kt
@@ -1,17 +1,22 @@
 package chat.rocket.android.main.presentation
 
+import chat.rocket.android.core.lifecycle.CancelStrategy
 import chat.rocket.android.infrastructure.LocalRepository
 import chat.rocket.android.server.domain.GetCurrentServerInteractor
 import chat.rocket.android.server.infraestructure.ConnectionManagerFactory
 import chat.rocket.android.server.infraestructure.RocketChatClientFactory
+import chat.rocket.android.util.extensions.launchUI
 import chat.rocket.common.RocketChatException
+import chat.rocket.common.util.ifNull
 import chat.rocket.core.RocketChatClient
 import chat.rocket.core.internal.realtime.disconnect
+import chat.rocket.core.internal.rest.logout
 import chat.rocket.core.internal.rest.unregisterPushToken
-import timber.log.Timber
 import javax.inject.Inject
 
-class MainPresenter @Inject constructor(private val navigator: MainNavigator,
+class MainPresenter @Inject constructor(private val view: MainView,
+                                        private val strategy: CancelStrategy,
+                                        private val navigator: MainNavigator,
                                         private val serverInteractor: GetCurrentServerInteractor,
                                         private val localRepository: LocalRepository,
                                         managerFactory: ConnectionManagerFactory,
@@ -30,19 +35,21 @@ class MainPresenter @Inject constructor(private val navigator: MainNavigator,
      * Logout from current server.
      */
     fun logout() {
-        // TODO: inject CancelStrategy, and MainView.
-//        launchUI(strategy) {
+        launchUI(strategy) {
             try {
-//                clearTokens()
-//                client.logout()
+                clearTokens()
+                client.logout()
                 //TODO: Add the code to unsubscribe to all subscriptions.
                 client.disconnect()
-//                view.onLogout()
-            } catch (e: RocketChatException) {
-                Timber.e(e)
-//                view.showMessage(e.message!!)
+                view.onLogout()
+            } catch (exception: RocketChatException) {
+                exception.message?.let {
+                    view.showMessage(it)
+                }.ifNull {
+                    view.showGenericErrorMessage()
+                }
             }
-//        }
+        }
     }
 
     private suspend fun clearTokens() {
