+++ app/src/main/java/chat/rocket/android/profile/di/ProfileFragmentModule.kt
@@ -1,13 +1,11 @@
 package chat.rocket.android.profile.di
 
 import android.arch.lifecycle.LifecycleOwner
-import chat.rocket.android.core.lifecycle.CancelStrategy
 import chat.rocket.android.dagger.scope.PerFragment
 import chat.rocket.android.profile.presentation.ProfileView
 import chat.rocket.android.profile.ui.ProfileFragment
 import dagger.Module
 import dagger.Provides
-import kotlinx.coroutines.experimental.Job
 
 @Module
 @PerFragment
@@ -21,10 +19,5 @@ class ProfileFragmentModule {
     @Provides
     fun provideLifecycleOwner(frag: ProfileFragment): LifecycleOwner {
         return frag
-    }
-
-    @Provides
-    fun provideCancelStrategy(owner: LifecycleOwner, jobs: Job): CancelStrategy {
-        return CancelStrategy(owner, jobs)
     }
 }
